feat: allow to cancel waiting approval issue in issue detail page
