+++ frontend/src/components/Issue/StatusTransitionButtonGroup/index.ts
@@ -1,0 +1,9 @@
+import CombinedRolloutButtonGroup from "./CombinedRolloutButtonGroup.vue";
+import StandaloneIssueStatusTransitionButtonGroup from "./StandaloneIssueStatusTransitionButtonGroup.vue";
+import TaskExtraActionsButton from "./TaskExtraActionsButton.vue";
+
+export {
+  StandaloneIssueStatusTransitionButtonGroup,
+  CombinedRolloutButtonGroup,
+  TaskExtraActionsButton,
+};
