+++ frontend/src/components/Issue/index.ts
@@ -1,7 +1,6 @@
 import IssueBanner from "./IssueBanner.vue";
 import IssueHighlightPanel from "./IssueHighlightPanel.vue";
 import IssueStagePanel from "./IssueStagePanel.vue";
-import IssueStatusTransitionButtonGroup from "./IssueStatusTransitionButtonGroup.vue";
 import IssueOutputPanel from "./IssueOutputPanel.vue";
 import IssueSidebar from "./IssueSidebar.vue";
 import IssueTaskStatementPanel from "./IssueTaskStatementPanel.vue";
@@ -15,7 +14,6 @@ export {
   IssueBanner,
   IssueHighlightPanel,
   IssueStagePanel,
-  IssueStatusTransitionButtonGroup,
   IssueOutputPanel,
   IssueSidebar,
   IssueTaskStatementPanel,
